chore(staking scripts): move to separate dir
