+++ scripts/EW/staking-snapshots/read-staking-snapshot-roles.ts
@@ -1,4 +1,4 @@
-import { chainConfigs, initWithPrivateKeySigner, NamespaceType } from '../../src';
+import { chainConfigs, initWithPrivateKeySigner, NamespaceType } from '../../../src';
 
 // CONFIG
 const CHAIN_ID = 246;
@@ -22,7 +22,13 @@ const rpcUrl = chainConfigs()[CHAIN_ID].rpcUrl;
   const { domainsService } = await connectToCacheServer();
 
   const domains = await domainsService.getRolesByNamespace({ parentType: NamespaceType.Application, namespace: namespace });
-  domains.forEach(d => {
-    console.log(d.definition)
-  })
+  const snap4Role = domains.filter(d => d.name == "snapshot4")[0];
+  const silverRole = domains.filter(d => d.name == "silverpool")[0];
+  const goldRole = domains.filter(d => d.name == "goldpool")[0];
+  console.log(snap4Role.definition);
+  console.log(silverRole.definition);
+  console.log(goldRole.definition);
+  // domains.forEach(d => {
+  //   console.log(d.definition)
+  // })
 })();
